model: Move the license view test classes into separate files
